+++ src/app/components/toolbar/toolbar-accelerometer.jsx
@@ -1,13 +1,12 @@
 'use strict';
 
 import React from 'react';
-import ToolbarSeparator from 'material-ui/Toolbar/ToolbarSeparator';
 import FontIcon from 'material-ui/FontIcon';
-import Paper from 'material-ui/Paper';
 import IconList from 'app/components/icon/icon-list';
 import AppPalette from 'app/configs/app-palette';
 import TextField from 'material-ui/TextField';
 import RaisedButton from 'material-ui/RaisedButton';
+import PanelToolbarBase from 'app/components/panel/panel-toolbar-base';
 
 const rotationSets = {
 	rotation0: {x: 0, y: 5.9, z: 0},
@@ -15,6 +14,32 @@ const rotationSets = {
 	rotation180: {x: 0, y: -5.9, z: 0},
 	rotation90: {x: 5.9, y: 0, z: 0} // 270 google
 };
+
+const buttons = [
+	{
+		id: 'rotation0',
+		tooltip: '0°',
+		fontIcon: 'mdi mdi-cellphone-android'
+	},
+	{
+		id: 'rotation90',
+		tooltip: '90°',
+		fontIcon: 'mdi mdi-cellphone-android',
+		iconStyle: {transform: 'rotate(90deg)'}
+	},
+	{
+		id: 'rotation180',
+		tooltip: '180°',
+		fontIcon: 'mdi mdi-cellphone-android',
+		iconStyle: {transform: 'rotate(180deg)'}
+	},
+	{
+		id: 'rotation270',
+		tooltip: '270°',
+		fontIcon: 'mdi mdi-cellphone-android',
+		iconStyle: {transform: 'rotate(270deg)'}
+	}
+];
 
 const ToolbarAccelerometer = class extends React.Component {
 	setRefX = c => {
@@ -59,45 +84,9 @@ const ToolbarAccelerometer = class extends React.Component {
 	}
 
 	render() {
-		const styles = {
-			button: {
-				marginTop: 5,
-				float: 'left'
-			},
-			items: {
-				float: 'left'
-			},
-			paper: {
-				height: 56
-			},
-			buttonSubmit: {
-				float: 'left',
-				marginTop: 10,
-				marginLeft: 15
-			},
-			icon: {
-				margin: '15px 10px 0 10px',
-				float: 'left'
-			},
-			sensorIcon: {
-				fontSize: '34px',
-				padding: '7px 11px 11px 0'
-			},
-			sensorIconRotationHorizontal: {
-				fontSize: '34px',
-				padding: '7px 11px 11px 0',
-				transform: 'rotate(-45deg)'
-			},
-			sensorIconRotationVertical: {
-				fontSize: '34px',
-				padding: '7px 11px 11px 0',
-				transform: 'rotate(45deg)'
-			},
-			sensorIconRotation: {}
-		};
-
 		const stylesInput = {
 			width: 100,
+			marginTop: -15,
 			float: 'left'
 		};
 
@@ -106,37 +95,6 @@ const ToolbarAccelerometer = class extends React.Component {
 			marginTop: 10,
 			marginLeft: 15
 		};
-
-		const styleSeparator = {
-			float: 'left',
-			margin: '0 0px 0 0px'
-		};
-
-		const buttons = [
-			{
-				id: 'rotation0',
-				tooltip: '0°',
-				fontIcon: 'mdi mdi-cellphone-android'
-			},
-			{
-				id: 'rotation90',
-				tooltip: '90°',
-				fontIcon: 'mdi mdi-cellphone-android',
-				iconStyle: {transform: 'rotate(90deg)'}
-			},
-			{
-				id: 'rotation180',
-				tooltip: '180°',
-				fontIcon: 'mdi mdi-cellphone-android',
-				iconStyle: {transform: 'rotate(180deg)'}
-			},
-			{
-				id: 'rotation270',
-				tooltip: '270°',
-				fontIcon: 'mdi mdi-cellphone-android',
-				iconStyle: {transform: 'rotate(270deg)'}
-			}
-		];
 
 		const props = {
 			buttons,
@@ -149,13 +107,11 @@ const ToolbarAccelerometer = class extends React.Component {
 		const renderedButtons = <IconList {...props}/>;
 
 		return (
-			<Paper style={Object.assign(this.props.style, styles.paper)} zDepth={1}>
-				<FontIcon style={styles.icon} className="mdi mdi-screen-rotation" color="rgba(0, 0, 0, 0.4)"/>
-				<ToolbarSeparator style={styleSeparator}/>
+			<PanelToolbarBase icon={<FontIcon className="mdi mdi-screen-rotation"/>}>
 				{renderedButtons}
-				<TextField name="fieldLiveSensorAccelerometerX" defaultValue={this.props.accelerometer.x} style={stylesInput} ref={this.setRefX} hintText="x" onFocus={this.props.onInputFocus} onBlur={this.props.onInputBlur}/>
-				<TextField name="fieldLiveSensorAccelerometerY" defaultValue={this.props.accelerometer.y} style={stylesInput} ref={this.setRefY} hintText="y" onFocus={this.props.onInputFocus} onBlur={this.props.onInputBlur}/>
-				<TextField name="fieldLiveSensorAccelerometerZ" defaultValue={this.props.accelerometer.z} style={stylesInput} ref={this.setRefZ} hintText="z" onFocus={this.props.onInputFocus} onBlur={this.props.onInputBlur}/>
+				<TextField name="fieldLiveSensorAccelerometerX" defaultValue={this.props.accelerometer.x} style={stylesInput} ref={this.setRefX} floatingLabelText="x" onFocus={this.props.onInputFocus} onBlur={this.props.onInputBlur}/>
+				<TextField name="fieldLiveSensorAccelerometerY" defaultValue={this.props.accelerometer.y} style={stylesInput} ref={this.setRefY} floatingLabelText="y" onFocus={this.props.onInputFocus} onBlur={this.props.onInputBlur}/>
+				<TextField name="fieldLiveSensorAccelerometerZ" defaultValue={this.props.accelerometer.z} style={stylesInput} ref={this.setRefZ} floatingLabelText="z" onFocus={this.props.onInputFocus} onBlur={this.props.onInputBlur}/>
 				<RaisedButton
 					className="btLiveSensorAccelerometerSubmit"
 					label="Submit"
@@ -164,23 +120,13 @@ const ToolbarAccelerometer = class extends React.Component {
 					onClick={this.handleClick}
 					style={styleButtonSubmit}
 					/>
-			</Paper>
+			</PanelToolbarBase>
 		);
 	}
 };
 
-ToolbarAccelerometer.contextTypes = {
-	muiTheme: React.PropTypes.object,
-	router: React.PropTypes.object
-};
-
-ToolbarAccelerometer.defaultProps = {
-	style: {}
-};
-
 ToolbarAccelerometer.propTypes = {
 	onClickBack: React.PropTypes.func,
-	style: React.PropTypes.object,
 	onChange: React.PropTypes.func,
 	onInputBlur: React.PropTypes.func,
 	onInputFocus: React.PropTypes.func,