+++ src/app/components/toolbar/toolbar-gps.jsx
@@ -1,95 +1,66 @@
 'use strict';
 
-// Vendor
 import React from 'react';
-import ToolbarSeparator from 'material-ui/Toolbar/ToolbarSeparator';
 import FontIcon from 'material-ui/FontIcon';
 import TextField from 'material-ui/TextField';
 import RaisedButton from 'material-ui/RaisedButton';
-import Paper from 'material-ui/Paper';
+import PanelToolbarBase from 'app/components/panel/panel-toolbar-base';
 
-// APP
 const ToolbarGPS = class extends React.Component {
-
-	constructor(props) {
-		super(props);
-		this.handleClickGPS = this.handleClickGPS.bind(this);
-		this.setRefLat = c => {
-			this.lat = c;
-		};
-		this.setRefLon = c => {
-			this.lon = c;
-		};
-	}
-
-	render() {
-		const styles = {
-			button: {
-				marginTop: 5,
-				float: 'left'
-			},
-			separator: {
-				margin: '0 15px 0 0px',
-				float: 'left'
-			},
-			items: {
-				float: 'left'
-			},
-			paper: {
-				height: 56
-			},
-			buttonSubmit: {
-				float: 'left',
-				marginTop: 10,
-				marginLeft: 15
-			},
-			icon: {
-				margin: '15px 10px 0 10px',
-				float: 'left'
-			}
-		};
-		return (
-			<Paper style={Object.assign(this.props.style, styles.paper)} zDepth={1}>
-				<FontIcon style={styles.icon} className="mdi mdi-map-marker" color="rgba(0, 0, 0, 0.4)"/>
-				<ToolbarSeparator style={styles.separator}/>
-				<TextField name="fieldLiveSensorGPSLatitude" style={styles.items} ref={this.setRefLat} hintText="latitude" onFocus={this.props.onInputFocus} onBlur={this.props.onInputBlur}/>
-				<TextField name="fieldLiveSensorGPSLongitude" style={styles.items} ref={this.setRefLon} hintText="longitude" onFocus={this.props.onInputFocus} onBlur={this.props.onInputBlur}/>
-				<RaisedButton
-					className="btLiveSensorGPSSubmit"
-					label="Submit"
-					title="Submit"
-					primary
-					onClick={this.handleClickGPS}
-					style={styles.buttonSubmit}
-					/>
-			</Paper>
-		);
-	}
-
-	handleClickGPS(e) {
+	handleClickGPS = e => {
 		const payload = {
 			latitude: parseFloat(this.lat.getValue()),
 			longitude: parseFloat(this.lon.getValue())
 		};
 		this.props.onChange(e, payload);
 	}
-};
 
-ToolbarGPS.contextTypes = {
-	muiTheme: React.PropTypes.object,
-	router: React.PropTypes.object
-};
+	setRefLat = c => {
+		this.lat = c;
+	}
 
-ToolbarGPS.defaultProps = {
-	style: {}
+	setRefLon = c => {
+		this.lon = c;
+	}
+
+	render() {
+		const styleButtonSubmit = {
+			float: 'left',
+			marginTop: 10,
+			marginLeft: 15
+		};
+		const styleLon = {
+			float: 'left',
+			marginTop: -15
+		};
+		const styleLat = {
+			float: 'left',
+			marginTop: -15,
+			marginLeft: 15
+		};
+		return (
+			<PanelToolbarBase icon={<FontIcon className="mdi mdi-map-marker"/>}>
+				<TextField defaultValue={this.props.gps.latitude} name="fieldLiveSensorGPSLatitude" style={styleLat} ref={this.setRefLat} floatingLabelText="latitude" onFocus={this.props.onInputFocus} onBlur={this.props.onInputBlur}/>
+				<TextField defaultValue={this.props.gps.longitude} name="fieldLiveSensorGPSLongitude" style={styleLon} ref={this.setRefLon} floatingLabelText="longitude" onFocus={this.props.onInputFocus} onBlur={this.props.onInputBlur}/>
+				<RaisedButton
+					className="btLiveSensorGPSSubmit"
+					label="Submit"
+					title="Submit"
+					primary
+					onClick={this.handleClickGPS}
+					style={styleButtonSubmit}
+					/>
+			</PanelToolbarBase>
+		);
+	}
 };
 
 ToolbarGPS.propTypes = {
 	onClickBack: React.PropTypes.func,
-	style: React.PropTypes.object,
 	onChange: React.PropTypes.func,
 	onInputBlur: React.PropTypes.func,
-	onInputFocus: React.PropTypes.func
+	onInputFocus: React.PropTypes.func,
+	gps: React.PropTypes.object
 };
 
 module.exports = ToolbarGPS;